refactor: move describe_link to policy
